chore(#40): refactor types
